+++ protector/helpers.py
@@ -1,3 +1,5 @@
+from past.builtins import basestring
+from functools import wraps
 from django.contrib.auth.models import Permission
 from django.contrib.auth import get_user_model
 from django.contrib.contenttypes.models import ContentType
@@ -10,8 +12,24 @@ from protector.internals import (
     _get_permission_filter, VIEW_RESTRICTED_OBJECTS, _get_permissions_query,
 )
 
+from protector.exceptions import NoReasonSpecified, ImproperResponsibleInstancePassed
+
 
 _view_perm = None
+
+
+def check_responsible_reason(func):
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        responsible = kwargs.get('responsible') or kwargs.get('responsible_id')
+        reason = kwargs.get('reason') or (len(args) > 2 and args[2])
+        if responsible is not None and not isinstance(responsible, get_user_model()):
+            raise ImproperResponsibleInstancePassed
+        if not isinstance(reason, basestring) or not len(reason):
+            raise NoReasonSpecified
+
+        return func(*args, **kwargs)
+    return wrapper
 
 
 def get_all_permission_owners(permission, include_superuser=False, obj=None):
